requirements_freeze.txt
fastapi
uvicorn
tqdm
librosa
pyaml
tensorboard
torchaudio==2.1.2
pesq
pystoi
torch-pesq
